all: fix misuses of math.Pow(x, 2)
math.Pow(x, 2) is really inefficient because math.Pow(x, y) is written
for the general use case of computing the power of x where y is a
floating point number.
In the simpler case of y as an integer, a much simpler algorithm can be
used.
In the even simpler case of y==2, inlining is even faster.